Show background jobs on Docker start
docker-sync prompts for upgrade sometimes during this command. This
leads to docker-sync failing to start and hanging there in background
with this prompt indefinitely.

This prompt is invisible, because two commands run and the same time and
override output from each other unpredictably.

Showing the list of background jobs at the end should make the problem
visible.